Fix encoder dtype handling; add proper bfloat16 support
NVIDIA's PR removed torch.autocast from the encoder without adding an
explicit feature cast, making float32 models slower (no tensor cores)
and making model.to(bfloat16) fail with a dtype mismatch error.

Fix: keep fp16 autocast for float32 models; cast features explicitly
to enc_dtype for models already in bfloat16/fp16.

from_pretrained() now accepts dtype=torch.bfloat16 to properly move
encoder/decoder/joint to bfloat16 while keeping the preprocessor in
float32 (required by torch.stft).

Also revert the NeMo .to(bfloat16) added to benchmark.py — that only
made NeMo look faster without touching our model, skewing the comparison.

+++ benchmark.py
@@ -54,7 +54,7 @@ def main():
 
     import nemo.collections.asr as nemo_asr
 
-    model = nemo_asr.models.EncDecRNNTBPEModel.from_pretrained(model_name=args.model).to(torch.bfloat16)
+    model = nemo_asr.models.EncDecRNNTBPEModel.from_pretrained(model_name=args.model)
     model = model.cuda()
     model.eval()
